Add Datechecker to goodreads API.
The book model now has methods defined which check the last updated date of the goodreads rating. If it it older than one day, the api to goodreads is called and resulting value stored in my Books table. If not, the value is pulled from the database without going to goodreads.